@@ -1,19 +1,19 @@
 {
-  "name": "portreact",
+  "name": "new-port",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "node-sass": "^4.11.0",
-    "re-base": "^4.0.0",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/user-event": "^7.1.2",
+    "node-sass": "^4.14.1",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-reveal": "^1.2.2",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.1.3",
-    "react-scroll": "^1.7.11",
-    "rxjs": "^6.3.3",
-    "sal.js": "^0.5.0",
-    "scroll-out": "^2.2.7"
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "3.4.3",
+    "react-scroll": "^1.8.1",
+    "scroll-out": "^2.2.12"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -24,10 +24,16 @@
   "eslintConfig": {
     "extends": "react-app"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
