@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from 'react';
 import {
 	BrowserRouter as Router,
 	Route,
@@ -12,22 +12,20 @@ import Home from "./pages/Home";
 //page not found
 import PageNotFound from "./layouts/PageNotFound";
 
-class App extends Component {
-	render() {
-		return (
-			<Router>
-				<div className="App">
-					<Navbar />
-					<Switch>
-						<Route exact path="/" component={Home} />
-						<Route path="/404" component={PageNotFound} />
-						<Redirect from="*" to="/404" />
-					</Switch>
-					<Footer />
-				</div>
-			</Router>
-		);
-	}
+function App() {
+  return (
+    <Router>
+      <div className="App">
+        <Navbar />
+        <Switch>
+          <Route exact path="/" component={Home} />
+          <Route path="/404" component={PageNotFound} />
+          <Redirect from="*" to="/404" />
+        </Switch>
+        <Footer />
+      </div>
+    </Router>
+  );
 }
 
 export default App;
